fixes / add Set and Select
Added in Select and Set functionality. Fixed the set ISR (connected to incorrect pin... now using digitalPinToInterrupt() to avoid confusion)
Discontinued use of delay in ISR (they do not work inside of ISR's)
Now when the select button is ON 1 it is for practice mode. otherwise it is listening for gunshot